pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pft-wallet"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "boto3",
    "dynaconf",
    "websockets",
    "web3",
    "click",
    "setuptools",
    # Task Node dependencies
    "tasknode-core",
    "tasknode-sdk",
    # XRPL dependencies
    "xrpl-py",
    "cryptography"
]

[tool.setuptools.package-data]
pft_wallet = ["static/**/*"]

[project.scripts]
pft-wallet = "pft_wallet.cli:main"

[tool.setuptools]
packages = ["pft_wallet"]
package-dir = {"" = "src"}

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy"
]